rename the package firebase
