+++ app/src/main/java/net/squanchy/eventdetails/EventDetailsModule.kt
@@ -1,6 +1,6 @@
 package net.squanchy.eventdetails
 
-import net.squanchy.service.firestore.FirebaseAuthService
+import net.squanchy.service.firebase.FirebaseAuthService
 
 import dagger.Module
 import dagger.Provides
