+++ app/src/main/java/net/squanchy/eventdetails/EventDetailsService.kt
@@ -5,7 +5,7 @@ import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Single
 import net.squanchy.schedule.domain.view.Event
-import net.squanchy.service.firestore.FirebaseAuthService
+import net.squanchy.service.firebase.FirebaseAuthService
 import net.squanchy.service.repository.EventRepository
 import net.squanchy.support.lang.Optional
