+++ app/src/main/java/net/squanchy/schedule/ScheduleService.kt
@@ -8,12 +8,12 @@ import net.squanchy.schedule.domain.view.Schedule
 import net.squanchy.schedule.domain.view.SchedulePage
 import net.squanchy.schedule.domain.view.Track
 import net.squanchy.schedule.tracksfilter.TracksFilter
-import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.firestore.model.schedule.FirestoreEvent
-import net.squanchy.service.firestore.model.schedule.FirestoreFavorite
-import net.squanchy.service.firestore.model.schedule.FirestoreSchedulePage
-import net.squanchy.service.firestore.toEvent
+import net.squanchy.service.firebase.FirebaseAuthService
+import net.squanchy.service.firebase.FirestoreDbService
+import net.squanchy.service.firebase.model.schedule.FirestoreEvent
+import net.squanchy.service.firebase.model.schedule.FirestoreFavorite
+import net.squanchy.service.firebase.model.schedule.FirestoreSchedulePage
+import net.squanchy.service.firebase.toEvent
 import net.squanchy.support.lang.Checksum
 import org.joda.time.DateTimeZone
 import org.joda.time.LocalDate
