+++ app/src/main/java/net/squanchy/search/SearchService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import net.squanchy.schedule.domain.view.Event;
 import net.squanchy.search.engines.SearchEngines;
-import net.squanchy.service.firestore.FirebaseAuthService;
+import net.squanchy.service.firebase.FirebaseAuthService;
 import net.squanchy.service.repository.EventRepository;
 import net.squanchy.service.repository.SpeakerRepository;
 import net.squanchy.speaker.domain.view.Speaker;
