+++ app/src/main/java/net/squanchy/service/firebase/FirebaseAuthService.kt
@@ -1,4 +1,4 @@
-package net.squanchy.service.firestore
+package net.squanchy.service.firebase
 
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.tasks.Task
