+++ app/src/main/java/net/squanchy/service/repository/firestore/FirestoreSpeakerRepository.kt
@@ -1,8 +1,8 @@
 package net.squanchy.service.repository.firestore
 
 import io.reactivex.Observable
-import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.firestore.toSpeaker
+import net.squanchy.service.firebase.FirestoreDbService
+import net.squanchy.service.firebase.toSpeaker
 import net.squanchy.service.repository.SpeakerRepository
 import net.squanchy.speaker.domain.view.Speaker
 import net.squanchy.support.lang.Checksum
