+++ app/src/main/java/net/squanchy/service/repository/firestore/FirestoreTracksRepository.kt
@@ -2,9 +2,9 @@ package net.squanchy.service.repository.firestore
 
 import io.reactivex.Observable
 import net.squanchy.schedule.domain.view.Track
-import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.firestore.model.schedule.FirestoreTrack
-import net.squanchy.service.firestore.toTrack
+import net.squanchy.service.firebase.FirestoreDbService
+import net.squanchy.service.firebase.model.schedule.FirestoreTrack
+import net.squanchy.service.firebase.toTrack
 import net.squanchy.service.repository.TracksRepository
 
 class FirestoreTracksRepository(private val firestoreDbService: FirestoreDbService) : TracksRepository {
