+++ app/src/main/java/net/squanchy/signin/SignInModule.kt
@@ -1,6 +1,6 @@
 package net.squanchy.signin
 
-import net.squanchy.service.firestore.FirebaseAuthService
+import net.squanchy.service.firebase.FirebaseAuthService
 
 import dagger.Module
 import dagger.Provides
