+++ app/src/main/java/net/squanchy/venue/VenueInfoService.kt
@@ -1,10 +1,10 @@
 package net.squanchy.venue
 
 import io.reactivex.Observable
-import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.firestore.model.conferenceinfo.FirestoreVenue
-import net.squanchy.service.firestore.toVenue
+import net.squanchy.service.firebase.FirebaseAuthService
+import net.squanchy.service.firebase.FirestoreDbService
+import net.squanchy.service.firebase.model.conferenceinfo.FirestoreVenue
+import net.squanchy.service.firebase.toVenue
 import net.squanchy.venue.domain.view.Venue
 
 internal class VenueInfoService(private val dbService: FirestoreDbService, private val authService: FirebaseAuthService) {
