+++ app/src/test/java/net/squanchy/schedule/FirestoreScheduleServiceTest.kt
@@ -9,14 +9,14 @@ import net.squanchy.schedule.domain.view.aSchedulePage
 import net.squanchy.schedule.domain.view.aTrack
 import net.squanchy.schedule.domain.view.anEvent
 import net.squanchy.schedule.tracksfilter.TracksFilter
-import net.squanchy.service.firestore.FirebaseAuthService
-import net.squanchy.service.firestore.FirestoreDbService
-import net.squanchy.service.firestore.aFirestoreDay
-import net.squanchy.service.firestore.aFirestoreEvent
-import net.squanchy.service.firestore.aFirestoreSchedulePage
-import net.squanchy.service.firestore.aFirestoreSpeaker
-import net.squanchy.service.firestore.aFirestoreTrack
-import net.squanchy.service.firestore.model.schedule.FirestoreFavorite
+import net.squanchy.service.firebase.FirebaseAuthService
+import net.squanchy.service.firebase.FirestoreDbService
+import net.squanchy.service.firebase.aFirestoreDay
+import net.squanchy.service.firebase.aFirestoreEvent
+import net.squanchy.service.firebase.aFirestoreSchedulePage
+import net.squanchy.service.firebase.aFirestoreSpeaker
+import net.squanchy.service.firebase.aFirestoreTrack
+import net.squanchy.service.firebase.model.schedule.FirestoreFavorite
 import net.squanchy.support.lang.Checksum
 import net.squanchy.support.lang.Optional
 import org.joda.time.LocalDate
